backend/requirements.txt
fastapi==0.114.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.6

geopy~=2.4.1
pyproj~=3.6.1